refactor: improve plan exercise resolution and input validation
Introduce `resolvePlanExercise` helper using new `findByPlanDayIdAndExerciseId`
repository method for consistent plan exercise lookup across CRUD operations.

Change DTO fields to `long` and add `toNonNegativeInt` validation converting
to `int` with range checks.

Add logging to controller update endpoint and debug print to delete endpoint.

Reorder imports and format long parameter lists.

+++ fitrbackend/src/main/java/com/example/fitrbackend/dto/CreatePlanDayExerciseRequest.java
@@ -5,9 +5,9 @@ import lombok.Getter;
 @Getter
 public class CreatePlanDayExerciseRequest {
     private long exerciseId;
-    private int targetSets;
-    private int targetReps;
-    private int targetDurationSeconds;
+    private long targetSets;
+    private long targetReps;
+    private long targetDurationSeconds;
     private float targetDistance;
     private float targetCalories;
     private Float targetWeight;
